Track permissions and options that are checked
These are used to provide tabcompletion for permission- and option-
using commands. This hooks into the same system used to notify for
debug mode, meaning that implementations pretty much get it
automatically. Perhaps at some point in the future a command could be
exposed to allow server admins to view a list of all known permissions.