Add initial Fabric implementation
Only has extremely rudimentary hooking for players -- most server
operations are not routed through permissions, but it's something to get
started.

Currently requires a customized variant of loom to build pending PR
acceptance.

IntelliJ doesn't like our usage of the RemapJarTask, so that section
might have to be commented out to correctly import this project.

there's a long way to go...