bukkit: Alow a fallback to op level check
This fallback will only happen if the permission is unset (value 0) in
PEX, and is configurable in a new platform-specific configuration
section.

The fallback happens through superperms and Vault checks, but not
through the native API -- the aim is to maintain expected results from
existing APIs without polluting PEX's core with knowledge of operator
status.

Closes #3016